Fix test (`aria-controls` wasn't specified in HTML). Improve API
Now API is more universal. The sheet body won't be automatically replaced with the component code.
Instead, the library user should do it personally.

`aria-controls` was set to `""` in HTML.

The test has been updated according to the API changes.
Additionally, `<template>` is now used to store sheet body
(so that it won't be shown to the user in case the page will be loading for a long time)